@@ -1,16 +1,14 @@
 import 'dart:io';
+import 'package:fab_circular_menu_plus/fab_circular_menu_plus.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:purchases_flutter/purchases_flutter.dart';
 import 'package:flutter_settings_screens/flutter_settings_screens.dart';
-import 'package:settings_ui/settings_ui.dart';
 import 'common_extension.dart';
-import 'common_widget.dart';
 import 'constant.dart';
-import 'plan_viewmodel.dart';
+import 'plan_provider.dart';
 import 'main.dart';
 import 'admob_banner.dart';
 
@@ -29,10 +27,20 @@ class MySettingsPage extends HookConsumerWidget {
     final isPremiumRestore = useState("premiumRestore".getSettingsValueBool(false));
     final premiumPrice = useState("premiumPrice".getSettingsValueString(""));
     final isReadError = useState(false);
-    final waitTime = useState("wait".getSettingsValueInt(waitTime_0));
-    final goTime = useState("go".getSettingsValueInt(goTime_0));
-    final flashTime = useState("flash".getSettingsValueInt(flashTime_0));
-    final isSound = useState('sound'.getSettingsValueBool(true));
+
+    final isSelectLeft  = useState(false);
+    final isSelectRight = useState(false);
+    final isSelectBack  = useState(false);
+
+    onSelectLeft() {
+      isSelectLeft.value = !isSelectLeft.value;
+      "isSelectLeft: ${isSelectLeft.value}".debugPrint();
+    }
+
+    onRightLeft() {
+      isSelectLeft.value = !isSelectLeft.value;
+      "isSelectRight: ${isSelectRight.value}".debugPrint();
+    }
 
     getPremiumPrice() async {
       final Offerings offerings = await Purchases.getOfferings();
@@ -42,156 +50,159 @@ class MySettingsPage extends HookConsumerWidget {
       }
     }
 
-    setTime(int time, String key) async {
-      await Settings.setValue('key_$key', time, notify: true);
-      ('${key}Time: $time').debugPrint();
-      if (key == "wait") waitTime.value = time;
-      if (key == "go") goTime.value = time;
-      if (key == "flash") flashTime.value = time;
-    }
+    final controller = useAnimationController(
+      duration: const Duration(milliseconds: 300),
+    );
+
+    final animation = Tween<double>(
+      begin: 0,
+      end: 1,
+    ).animate(controller);
 
     useEffect(() {
       WidgetsBinding.instance.addPostFrameCallback((_) async {
         initSettings();
-        if (!isPremiumProvider) {
-          await Purchases.setLogLevel(LogLevel.debug);
-          await Purchases.configure(PurchasesConfiguration(apiKey));
-          await Purchases.enableAdServicesAttributionTokenCollection();
-          Purchases.addReadyForPromotedProductPurchaseListener((productID, startPurchase) async {
-            'productID: $productID'.debugPrint();
-            try {
-              final purchaseResult = await startPurchase.call();
-              'productID: ${purchaseResult.productIdentifier}'.debugPrint();
-              'customerInfo: ${purchaseResult.customerInfo}'.debugPrint();
-            } on PlatformException catch (e) {
-              'Error: ${e.message}'.debugPrint();
-            }
-          });
-          if (premiumPrice.value == "") {
-            try {
-              getPremiumPrice();
-            } on PlatformException catch (e) {
-              isReadError.value = true;
-              'ReadError: ${isReadError.value}, Error: ${e.message}'.debugPrint();
-            }
-          }
-        }
+        // if (!isPremiumProvider) {
+        //   await Purchases.setLogLevel(LogLevel.debug);
+        //   await Purchases.configure(PurchasesConfiguration(apiKey));
+        //   await Purchases.enableAdServicesAttributionTokenCollection();
+        //   Purchases.addReadyForPromotedProductPurchaseListener((productID, startPurchase) async {
+        //     'productID: $productID'.debugPrint();
+        //     try {
+        //       final purchaseResult = await startPurchase.call();
+        //       'productID: ${purchaseResult.productIdentifier}'.debugPrint();
+        //       'customerInfo: ${purchaseResult.customerInfo}'.debugPrint();
+        //     } on PlatformException catch (e) {
+        //       'Error: ${e.message}'.debugPrint();
+        //     }
+        //   });
+        //   if (premiumPrice.value == "") {
+        //     try {
+        //       getPremiumPrice();
+        //     } on PlatformException catch (e) {
+        //       isReadError.value = true;
+        //       'ReadError: ${isReadError.value}, Error: ${e.message}'.debugPrint();
+        //     }
+        //   }
+        // }
       });
       "width: $width, height: $height".debugPrint();
-      "waitTIme: ${waitTime.value}, goTime: ${goTime.value}, flashTime: ${flashTime.value}".debugPrint();
       "isPremiumProvider: $isPremiumProvider, isPremium: ${isPremium.value}, isPremiumRestore: ${isPremiumRestore.value}".debugPrint();
       return;
     }, const []);
 
     return Scaffold(
-      appBar: settingsAppBar(context, context.settingsTitle(), false),
       body: Column(children: [
-        Flexible(child:
-          SettingsList(sections: [
-            ///Waiting Time
-            SettingsSection(title: Text(context.timeSettings()),
-              tiles: [
-                CustomSettingsTile(
-                  child: Container(
-                    padding: settingsTitlePadding(false),
-                    decoration: settingsTileDecoration(true, false),
-                    child: Column(crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        settingsTitle(context, context.waitTime(), waitTime.value),
-                        SliderTheme(
-                          data: sliderTheme(context, transpRedColor),
-                          child: Slider(
-                            value: waitTime.value.toDouble(),
-                            max: maxTime.toDouble(),
-                            min: minTime.toDouble(),
-                            divisions: maxTime - minTime + 1,
-                            onChanged: (double value) => setTime(value.toInt(), 'wait')
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                ///Going Time
-                CustomSettingsTile(
-                  child: Container(
-                    padding: settingsTitlePadding(false),
-                    color: (Platform.isIOS || Platform.isMacOS) ? whiteColor: transpColor,
-                    child: Column(crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        settingsTitle(context, context.goTime(), goTime.value),
-                        SliderTheme(
-                          data: sliderTheme(context, transpGreenColor),
-                          child: Slider(
-                            value: goTime.value.toDouble(),
-                            max: maxTime.toDouble(),
-                            min: minTime.toDouble(),
-                            divisions: maxTime - minTime + 1,
-                            onChanged: (double value) => setTime(value.toInt(), 'go')
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-                ///Flashing Time
-                CustomSettingsTile(
-                  child: Container(
-                    padding: settingsTitlePadding(true),
-                    decoration: settingsTileDecoration(false, true),
-                    child: Column(crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        settingsTitle(context, context.flashTime(), flashTime.value),
-                        SliderTheme(
-                          data: sliderTheme(context, transpYellowColor),
-                          child: Slider(
-                            value: flashTime.value.toDouble(),
-                            max: maxTime.toDouble(),
-                            min: minTime.toDouble(),
-                            divisions: maxTime - minTime + 1,
-                            onChanged: (double value) => setTime(value.toInt(), 'flash')
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ),
-              ]
-            ),
-            ///Sound On/Off
-            SettingsSection(
-              title: Text(context.soundSettings()),
-              tiles: [
-                SettingsTile.switchTile(
-                  leading: const Icon(Icons.music_note),
-                  activeSwitchColor: transpGreenColor,
-                  title: Text(context.crosswalkSound()),
-                  initialValue: isSound.value,
-                  onToggle: (value) async {
-                    await Settings.setValue<bool>('key_sound', value, notify: true);
-                    ('sound: $value').debugPrint();
-                    isSound.value = value;
-                  },
-                )
-              ]
-            ),
-            ///Premium Plan
-            if (!isPremiumProvider) SettingsSection(
-              title: Text(context.upgrade()),
-              tiles: [
-                SettingsTile(
-                  title: Text(context.settingsPremiumTitle(premiumPrice.value, isReadError.value)),
-                  leading: premiumPrice.value.settingsPremiumLeadingIcon(isReadError.value),
-                  trailing: premiumPrice.value.settingsPremiumTrailingIcon(),
-                  onPressed: (context) => (premiumPrice.value != "") ? context.pushUpgradePage(): null,
-                ),
-              ]
-            ),
-          ]),
-        ),
+        Row(children: [
+          SizedBox(width: context.height() * 0.05),
+          IconButton(
+            icon: const Icon(Icons.arrow_back_ios),
+            onPressed: () async => context.pushHomePage(),
+          ),
+          const Spacer(),
+        ]),
+        const Spacer(),
+        // Stack(
+        //   alignment: Alignment.center,
+        //   children: [
+        //     if (!isSelectLeft.value) FloatingActionButton(
+        //       shape: const CircleBorder(),
+        //       backgroundColor: blackColor,
+        //       heroTag: "select_left",
+        //       onPressed: () => onSelectLeft(),
+        //       child: const Icon(Icons.arrow_back),  // ユニークなタグを割り当てます
+        //     ),
+        //   ],
+        // ),
+        // const Spacer(),
+        // Flexible(child:
+        //   SettingsList(sections: [
+        //     ///Premium Plan
+        //     if (!isPremiumProvider) SettingsSection(
+        //       title: Text(context.upgrade()),
+        //       tiles: [
+        //         SettingsTile(
+        //           title: Text(context.settingsPremiumTitle(premiumPrice.value, isReadError.value)),
+        //           leading: premiumPrice.value.settingsPremiumLeadingIcon(isReadError.value),
+        //           trailing: premiumPrice.value.settingsPremiumTrailingIcon(),
+        //           onPressed: (context) => (premiumPrice.value != "") ? context.pushUpgradePage(): null,
+        //         ),
+        //       ]
+        //     ),
+        //   ]),
+        // ),
         ///AdMob Banner
         if (!isPremiumProvider) const AdBannerWidget()
-      ])
+      ]),
+      floatingActionButton: FabCircularMenuPlus(
+          alignment: Alignment.center,
+          ringWidth: context.height() * 0.15,
+          ringDiameter: context.height() * 0.6,
+          ringColor: yellowColor,
+          fabCloseColor: yellowColor,
+          fabOpenColor: whiteColor,
+          fabOpenIcon: const Icon(Icons.menu, color: blackColor),
+          fabCloseIcon: const Icon(Icons.close, color: blackColor),
+          children: [
+            FabCircularMenuPlus(
+                alignment: Alignment.center,
+                ringWidth: context.height() * 0.15,
+                ringDiameter: context.height() * 0.6,
+                ringColor: yellowColor,
+                fabCloseColor: yellowColor,
+                fabOpenColor: whiteColor,
+                fabOpenIcon: const Icon(Icons.menu, color: blackColor),
+                fabCloseIcon: const Icon(Icons.close, color: blackColor),
+                children: [
+                  GestureDetector(
+                      child: SizedBox(
+                        width: context.height() * 0.15,
+                        child: Image.asset(flagImageJP),
+                      ),
+                      onTap: () {
+                        'flag_jp'.debugPrint();
+                      }
+                  ),
+                  GestureDetector(
+                      child: SizedBox(
+                        width: context.height() * 0.15,
+                        child: Image.asset(flagImageUS),
+                      ),
+                      onTap: () {
+                        'flag_us'.debugPrint();
+                      }
+                  ),
+                  GestureDetector(
+                      child: SizedBox(
+                        width: context.height() * 0.15,
+                        child: Image.asset(flagImageUK),
+                      ),
+                      onTap: () {
+                        'flag_uk'.debugPrint();
+                      }
+                  ),
+                ]
+            ),
+            GestureDetector(
+              child: SizedBox(
+                width: context.height() * 0.15,
+                child: Image.asset(flagImageUS),
+              ),
+              onTap: () {
+                'flag_us'.debugPrint();
+              }
+            ),
+            GestureDetector(
+              child: SizedBox(
+                width: context.height() * 0.15,
+                child: Image.asset(flagImageUK),
+              ),
+              onTap: () {
+                'flag_uk'.debugPrint();
+              }
+            ),
+          ]
+      ),
     );
   }
 }
